avail config
