+++ crates/client/data-availability/Cargo.toml
@@ -12,6 +12,7 @@ repository = "https://github.com/keep-starknet-strange/madara"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+anyhow = { workspace = true }
 async-trait = { workspace = true }
 futures = "0.3.21"
 jsonrpsee = { version = "0.18.2", features = ["http-client", "ws-client", "macros"] }
@@ -42,8 +43,6 @@ ethers = "2.0.7"
 # Avail subxt dependency
 avail-subxt = { git="https://github.com/availproject/avail", version = "0.3.0", branch = "develop" }
 subxt = "0.29"
-anyhow = { workspace = true }
-sp-keyring = "*"
 
 # Celestia
 celestia-rpc = { git = "https://github.com/eigerco/celestia-node-rs" }
