+++ crates/client/data-availability/src/lib.rs
@@ -1,5 +1,7 @@
-pub mod celestia;
 pub mod ethereum;
+pub mod celestia;
+pub mod avail;
+pub mod utils;
 mod sharp;
 
 use std::collections::HashMap;
@@ -13,30 +15,15 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use ethers::types::U256;
 use futures::StreamExt;
-use lazy_static::lazy_static;
 use log::{debug, error, info, trace, warn};
-use mp_starknet::storage::{
-    PALLET_STARKNET, STARKNET_CONTRACT_CLASS, STARKNET_CONTRACT_CLASS_HASH, STARKNET_NONCE, STARKNET_STORAGE,
-};
+
 use sc_client_api::client::BlockchainEvents;
 use sp_api::ProvideRuntimeApi;
-use sp_io::hashing::twox_128;
 use sp_runtime::traits::Block as BlockT;
 use uuid::Uuid;
 
-const LOG_TARGET: &str = "da-worker";
-
-lazy_static! {
-    static ref SN_NONCE_PREFIX: Vec<u8> = [twox_128(PALLET_STARKNET), twox_128(STARKNET_NONCE)].concat();
-    static ref SN_CONTRACT_CLASS_HASH_PREFIX: Vec<u8> =
-        [twox_128(PALLET_STARKNET), twox_128(STARKNET_CONTRACT_CLASS_HASH)].concat();
-    static ref SN_CONTRACT_CLASS_PREFIX: Vec<u8> =
-        [twox_128(PALLET_STARKNET), twox_128(STARKNET_CONTRACT_CLASS)].concat();
-    static ref SN_STORAGE_PREFIX: Vec<u8> = [twox_128(PALLET_STARKNET), twox_128(STARKNET_STORAGE)].concat();
-}
-
-pub type StorageWrites<'a> = Vec<(&'a [u8], &'a [u8])>;
 pub struct DataAvailabilityWorker<B, C>(PhantomData<(B, C)>);
+
 
 #[async_trait]
 pub trait DaClient {
@@ -88,7 +75,7 @@ where
                 }
             }
 
-            let state_diff = pre_0_11_0_state_diff(storage_diffs, nonces);
+            let state_diff = utisl.pre_0_11_0_state_diff(storage_diffs, nonces);
 
             // Store the DA output from the SN OS
             if let Err(db_err) = madara_backend.da().store_state_diff(&storage_event.block, state_diff) {
@@ -159,27 +146,4 @@ where
             // }
         }
     }
-}
-
-// encode calldata:
-// - https://docs.starknet.io/documentation/architecture_and_concepts/Data_Availability/on-chain-data/#pre_v0.11.0_example
-pub fn pre_0_11_0_state_diff(storage_diffs: HashMap<&[u8], StorageWrites>, nonces: HashMap<&[u8], &[u8]>) -> Vec<U256> {
-    let mut state_diff: Vec<U256> = Vec::new();
-
-    state_diff.push(U256::from(storage_diffs.len()));
-
-    for (addr, writes) in storage_diffs {
-        state_diff.push(U256::from_big_endian(addr));
-        state_diff.push(U256::from(writes.len()));
-        for write in writes {
-            state_diff.push(U256::from_big_endian(write.0));
-            state_diff.push(U256::from_big_endian(write.1));
-        }
-    }
-
-    for (addr, nonce) in nonces {
-        state_diff.push(U256::from_big_endian(addr));
-        state_diff.push(U256::from_big_endian(nonce));
-    }
-    state_diff
 }
