+++ crates/node/src/command.rs
@@ -6,7 +6,7 @@ use pallet_starknet::utils;
 use sc_cli::{ChainSpec, RpcMethods, RuntimeVersion, SubstrateCli};
 
 use crate::benchmarking::{inherent_benchmark_data, RemarkBuilder};
-use crate::cli::{Cli, DaMode, Subcommand, Testnet};
+use crate::cli::{Cli, DaLayer, Subcommand, Testnet};
 use crate::{chain_spec, constants, service};
 impl SubstrateCli for Cli {
     fn impl_name() -> String {
@@ -275,7 +275,7 @@ pub fn run() -> sc_cli::Result<()> {
                 cli.run.run_cmd.rpc_methods = RpcMethods::Unsafe;
             }
 
-            let mut da_client: Option<(DaMode, PathBuf)> = None;
+            let mut da_client: Option<(DaLayer, PathBuf)> = None;
             if let Some(da_mode) = cli.run.da_mode.clone() {
                 let da_path = std::path::PathBuf::from(madara_path.clone() + "/da-" + da_mode.as_str() + ".json");
                 if !da_path.exists() {
